pyproject.toml
[project]
name = "avatar-mcp"
version = "0.1.0"
description = "Desktop avatar companion MCP server for Claude Code"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0.0",
    "PyQt6>=6.6.0",
    "edge-tts>=6.1.0",
    "kokoro-onnx>=0.5.0",
    "soundfile>=0.12.0",
    "SpeechRecognition>=3.10.0",
    "PyAudio>=0.2.14",
    "pygame>=2.5.0",
    "pyperclip>=1.8.0",
    "pyautogui>=0.9.54",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
premium = ["elevenlabs>=1.0.0"]
japanese = ["misaki[ja]"]
dev = ["pytest>=7.0.0"]

[project.scripts]
avatar-mcp = "avatar_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/avatar_mcp"]
